refactor: (BO) extract all components from Ingestion page
