+++ lib/zaq_web/components/bo_layout.ex
@@ -5,6 +5,9 @@ defmodule ZaqWeb.Components.BOLayout do
   Back office layout with collapsible sidebar and section dropdowns.
   """
   use Phoenix.Component
+
+  import ZaqWeb.CoreComponents, only: [theme_toggle: 1]
+
   alias Zaq.Addons.FeatureStore
   alias Zaq.System
   use ZaqWeb, :verified_routes
@@ -208,14 +211,14 @@ defmodule ZaqWeb.Components.BOLayout do
                   border-left: none;
                 }
       </style>
-      
+
     <!-- Sidebar -->
       <aside
         id="bo-sidebar"
         data-section-ids={Enum.join(@nav_section_ids, ",")}
         class="fixed top-0 left-0 h-screen zaq-sidebar flex flex-col z-40 shadow-xl"
       >
-        
+
     <!-- Logo + collapse toggle -->
         <div class="h-16 flex items-center justify-between px-3 flex-shrink-0 zaq-sidebar-header">
           <div class="flex items-center gap-2 min-w-0">
@@ -248,10 +251,10 @@ defmodule ZaqWeb.Components.BOLayout do
             </svg>
           </button>
         </div>
-        
+
     <!-- Nav -->
         <nav class="flex-1 py-3 px-2 overflow-y-auto overflow-x-hidden space-y-0.5">
-          
+
     <!-- Dashboard (standalone) -->
           <div class="nav-item-wrap">
             <a
@@ -281,7 +284,7 @@ defmodule ZaqWeb.Components.BOLayout do
             </a>
             <div class="nav-tooltip zaq-text-body-sm zaq-sidebar-nav-tooltip">Dashboard</div>
           </div>
-          
+
     <!-- Sections -->
           <%= for section <- @nav_sections do %>
             <.nav_section
@@ -352,7 +355,7 @@ defmodule ZaqWeb.Components.BOLayout do
           </div>
         </div>
       </aside>
-      
+
     <!-- Main -->
       <main id="bo-main" class="flex-1">
         <!-- Header -->
@@ -373,6 +376,7 @@ defmodule ZaqWeb.Components.BOLayout do
           </div>
 
           <div class="flex items-center gap-2 shrink-0">
+            <.theme_toggle />
             <a
               id="header-notifications-link"
               href={~p"/bo/channels/notifications/logs"}
