+++ lib/zaq_web/live/bo/ai/ingestion_live.html.heex
@@ -7,27 +7,7 @@
   features_version={@features_version}
   auto_dismiss={false}
 >
-  <div :if={not @embedding_ready} class="zaq-feedback-banner zaq-warning zaq-text-body">
-    <span class="zaq-feedback-icon">
-      <svg fill="none" stroke="currentColor" viewBox="0 0 24 24">
-        <path
-          stroke-linecap="round"
-          stroke-linejoin="round"
-          stroke-width="2"
-          d="M12 9v2m0 4h.01M10.29 3.86L1.82 18a2 2 0 001.71 3h16.94a2 2 0 001.71-3L13.71 3.86a2 2 0 00-3.42 0z"
-        />
-      </svg>
-    </span>
-    <div class="zaq-feedback-body">
-      <p class="font-semibold" data-testid="embedding-warning-title">Embedding not configured</p>
-      <p>
-        Please configure and save your embedding settings before ingesting documents.
-        <a href="/bo/system-config?tab=embedding" class="zaq-link-underline">
-          Go to Settings →
-        </a>
-      </p>
-    </div>
-  </div>
+  <.ingestion_embedding_banner :if={not @embedding_ready} />
 
   <div class="grid grid-cols-1 lg:grid-cols-3 gap-6">
     <%!-- =============== LEFT: File Browser + Upload =============== --%>
