+++ storybook/components/design_system/.index.exs
@@ -1,0 +1,6 @@
+defmodule Storybook.Components.DesignSystem do
+  use PhoenixStorybook.Index
+
+  def folder_name, do: "Design system"
+  def folder_index, do: 10
+end
