+++ storybook/components/design_system/modal_add_raw.story.exs
@@ -1,0 +1,16 @@
+defmodule Storybook.Components.DesignSystem.ModalAddRaw do
+  use PhoenixStorybook.Story, :component
+
+  def function, do: &ZaqWeb.Components.DesignSystem.ModalAddRaw.modal_add_raw/1
+  def description, do: "Add raw Markdown file modal."
+
+  def variations do
+    [
+      %Variation{
+        id: :default,
+        description: "Default",
+        attributes: %{modal_error: nil, modal_name: "my-note", current_dir: "docs"}
+      }
+    ]
+  end
+end
