+++ storybook/components/design_system/modal_new_folder.story.exs
@@ -1,0 +1,16 @@
+defmodule Storybook.Components.DesignSystem.ModalNewFolder do
+  use PhoenixStorybook.Story, :component
+
+  def function, do: &ZaqWeb.Components.DesignSystem.ModalNewFolder.modal_new_folder/1
+  def description, do: "Create new folder modal."
+
+  def variations do
+    [
+      %Variation{
+        id: :default,
+        description: "Default",
+        attributes: %{modal_error: nil, modal_name: "new-folder"}
+      }
+    ]
+  end
+end
