+++ test/e2e/support/story-urls.json
@@ -5,6 +5,24 @@
   "/storybook/chat_message/user_bubble",
   "/storybook/components/chat/chat_message",
   "/storybook/components/data/table",
+  "/storybook/components/design_system/breadcrumb",
+  "/storybook/components/design_system/dropzone",
+  "/storybook/components/design_system/ingestion_embedding_banner",
+  "/storybook/components/design_system/ingestion_file_browser_header",
+  "/storybook/components/design_system/ingestion_file_grid_view",
+  "/storybook/components/design_system/ingestion_file_icon",
+  "/storybook/components/design_system/ingestion_file_list_view",
+  "/storybook/components/design_system/ingestion_jobs_panel",
+  "/storybook/components/design_system/ingestion_volume_selector",
+  "/storybook/components/design_system/modal_add_raw",
+  "/storybook/components/design_system/modal_delete",
+  "/storybook/components/design_system/modal_delete_selected",
+  "/storybook/components/design_system/modal_new_folder",
+  "/storybook/components/design_system/modal_rename",
+  "/storybook/components/design_system/modal_move",
+  "/storybook/components/design_system/modal_share",
+  "/storybook/components/design_system/status_pill",
+  "/storybook/components/design_system/toggle",
   "/storybook/components/feedback/flash",
   "/storybook/components/feedback/status_badge",
   "/storybook/components/file_preview/file_preview",
